backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
docker
GitPython
